More changes in the writing of high score files. Not yet completed. Also minor bug fixed

+++ source/core/src/main/com/deco2800/game/components/scores/CalcScore.java
@@ -10,7 +10,7 @@ public class CalcScore {
      * @param clock - in seconds
      * @return A score which is proportional to the clocks time
      */
-    public long calculateScore(int clock) {
+    public int calculateScore(int clock) {
         int levelMaxTime = 3 * 60; // The player has to completed the level within 3 minutes
         int levelScore = 100; // The player will be awarded of 100 points per remaining second
         int score; // The players calculated score
